+++ modules/apps/web-experience/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/display/context/LayoutsAdminDisplayContext.java
@@ -16,7 +16,10 @@ package com.liferay.layout.admin.web.internal.display.context;
 
 import com.liferay.exportimport.kernel.staging.LayoutStagingUtil;
 import com.liferay.exportimport.kernel.staging.StagingUtil;
+import com.liferay.layout.admin.web.configuration.LayoutAdminWebConfiguration;
+import com.liferay.layout.admin.web.constants.LayoutAdminDisplayStyleKeys;
 import com.liferay.layout.admin.web.internal.constants.LayoutAdminPortletKeys;
+import com.liferay.layout.admin.web.internal.constants.LayoutAdminWebKeys;
 import com.liferay.layout.util.comparator.LayoutCreateDateComparator;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.dao.search.EmptyOnClickRowChecker;
@@ -33,6 +36,7 @@ import com.liferay.portal.kernel.model.LayoutRevision;
 import com.liferay.portal.kernel.model.LayoutSet;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
+import com.liferay.portal.kernel.portlet.LiferayWindowState;
 import com.liferay.portal.kernel.portlet.PortalPreferences;
 import com.liferay.portal.kernel.portlet.PortletPreferencesFactoryUtil;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
@@ -41,6 +45,8 @@ import com.liferay.portal.kernel.service.LayoutSetLocalServiceUtil;
 import com.liferay.portal.kernel.service.permission.GroupPermissionUtil;
 import com.liferay.portal.kernel.service.permission.LayoutPermissionUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.HtmlUtil;
+import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.OrderByComparator;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
@@ -52,13 +58,14 @@ import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.util.LayoutDescription;
 import com.liferay.portal.util.LayoutListUtil;
 import com.liferay.portlet.layoutsadmin.display.context.GroupDisplayContextHelper;
+import com.liferay.taglib.security.PermissionsURLTag;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
 import java.util.Objects;
 
-import javax.portlet.PortletRequest;
+import javax.portlet.ActionRequest;
 import javax.portlet.PortletURL;
 
 /**
@@ -83,34 +90,28 @@ public class LayoutsAdminDisplayContext {
 			WebKeys.LAYOUT_DESCRIPTIONS, getLayoutDescriptions());
 	}
 
-	public PortletURL getAddLayoutURL() {
+	public String getAddLayoutURL() {
 		return getAddLayoutURL(LayoutConstants.DEFAULT_PLID, isPrivatePages());
 	}
 
-	public PortletURL getAddLayoutURL(long selPlid, Boolean privateLayout) {
-		PortletURL addLayoutURL = PortalUtil.getControlPanelPortletURL(
-			_liferayPortletRequest, LayoutAdminPortletKeys.GROUP_PAGES,
-			PortletRequest.RENDER_PHASE);
+	public String getAddLayoutURL(long selPlid, Boolean privateLayout) {
+		PortletURL addLayoutURL = _liferayPortletResponse.createRenderURL();
 
 		addLayoutURL.setParameter("mvcPath", "/add_layout.jsp");
+		addLayoutURL.setParameter("redirect", _themeDisplay.getURLCurrent());
+		addLayoutURL.setParameter("backURL", _themeDisplay.getURLCurrent());
+		addLayoutURL.setParameter("groupId", String.valueOf(getSelGroupId()));
 
 		if (selPlid >= LayoutConstants.DEFAULT_PLID) {
 			addLayoutURL.setParameter("selPlid", String.valueOf(selPlid));
 		}
-
-		addLayoutURL.setParameter("groupId", String.valueOf(getSelGroupId()));
 
 		if (privateLayout != null) {
 			addLayoutURL.setParameter(
 				"privateLayout", String.valueOf(privateLayout));
 		}
 
-		addLayoutURL.setParameter(
-			"backURL",
-			PortalUtil.getCurrentURL(
-				PortalUtil.getHttpServletRequest(_liferayPortletRequest)));
-
-		return addLayoutURL;
+		return addLayoutURL.toString();
 	}
 
 	public JSONArray getBreadcrumbEntriesJSONArray() throws PortalException {
@@ -127,6 +128,10 @@ public class LayoutsAdminDisplayContext {
 		}
 
 		Layout selLayout = getSelLayout();
+
+		if (selLayout == null) {
+			return breadcrumbEntriesJSONArray;
+		}
 
 		List<Layout> layouts = selLayout.getAncestors();
 
@@ -147,6 +152,56 @@ public class LayoutsAdminDisplayContext {
 		return breadcrumbEntriesJSONArray;
 	}
 
+	public String getConfigureLayoutURL(Layout layout) {
+		PortletURL configureLayoutURL =
+			_liferayPortletResponse.createRenderURL();
+
+		configureLayoutURL.setParameter("mvcPath", "/edit_layout.jsp");
+		configureLayoutURL.setParameter(
+			"redirect", _themeDisplay.getURLCurrent());
+		configureLayoutURL.setParameter(
+			"backURL", _themeDisplay.getURLCurrent());
+		configureLayoutURL.setParameter(
+			"groupId", String.valueOf(layout.getGroupId()));
+		configureLayoutURL.setParameter(
+			"selPlid", String.valueOf(layout.getPlid()));
+		configureLayoutURL.setParameter(
+			"privateLayout", String.valueOf(layout.isPrivateLayout()));
+
+		return configureLayoutURL.toString();
+	}
+
+	public String getCopyApplicationsURL(Layout layout) {
+		PortletURL copyApplicationsURL =
+			_liferayPortletResponse.createRenderURL();
+
+		copyApplicationsURL.setParameter("mvcPath", "/copy_applications.jsp");
+		copyApplicationsURL.setParameter(
+			"redirect", _themeDisplay.getURLCurrent());
+		copyApplicationsURL.setParameter(
+			"backURL", _themeDisplay.getURLCurrent());
+		copyApplicationsURL.setParameter(
+			"groupId", String.valueOf(layout.getGroupId()));
+		copyApplicationsURL.setParameter(
+			"selPlid", String.valueOf(layout.getPlid()));
+		copyApplicationsURL.setParameter(
+			"privateLayout", String.valueOf(layout.isPrivateLayout()));
+
+		return copyApplicationsURL.toString();
+	}
+
+	public String getDeleteLayoutURL(Layout layout) {
+		PortletURL deleteLayoutURL = _liferayPortletResponse.createActionURL();
+
+		deleteLayoutURL.setParameter(
+			ActionRequest.ACTION_NAME, "/layout/delete_layout");
+		deleteLayoutURL.setParameter("redirect", _themeDisplay.getURLCurrent());
+		deleteLayoutURL.setParameter(
+			"selPlid", String.valueOf(layout.getPlid()));
+
+		return deleteLayoutURL.toString();
+	}
+
 	public String getDisplayStyle() {
 		if (Validator.isNotNull(_displayStyle)) {
 			return _displayStyle;
@@ -162,6 +217,10 @@ public class LayoutsAdminDisplayContext {
 		return _displayStyle;
 	}
 
+	public String getEditLayoutURL(Layout layout) throws PortalException {
+		return PortalUtil.getLayoutFullURL(layout, _themeDisplay);
+	}
+
 	public Group getGroup() {
 		return _groupDisplayContextHelper.getGroup();
 	}
@@ -175,29 +234,33 @@ public class LayoutsAdminDisplayContext {
 	}
 
 	public JSONArray getLayoutColumnsJSONArray() throws Exception {
-		JSONArray layoutBlocksJSONArray = JSONFactoryUtil.createJSONArray();
+		JSONArray layoutColumnsJSONArray = JSONFactoryUtil.createJSONArray();
 
-		layoutBlocksJSONArray.put(_getLayoutsJSONArray(0));
+		layoutColumnsJSONArray.put(_getLayoutsJSONArray(0));
 
 		if (getSelPlid() == LayoutConstants.DEFAULT_PLID) {
-			return layoutBlocksJSONArray;
+			return layoutColumnsJSONArray;
 		}
 
 		Layout selLayout = getSelLayout();
+
+		if (selLayout == null) {
+			return layoutColumnsJSONArray;
+		}
 
 		List<Layout> layouts = selLayout.getAncestors();
 
 		Collections.reverse(layouts);
 
 		for (Layout layout : layouts) {
-			layoutBlocksJSONArray.put(
+			layoutColumnsJSONArray.put(
 				_getLayoutsJSONArray(layout.getLayoutId()));
 		}
 
-		layoutBlocksJSONArray.put(
+		layoutColumnsJSONArray.put(
 			_getLayoutsJSONArray(selLayout.getLayoutId()));
 
-		return layoutBlocksJSONArray;
+		return layoutColumnsJSONArray;
 	}
 
 	public List<LayoutDescription> getLayoutDescriptions() {
@@ -343,6 +406,19 @@ public class LayoutsAdminDisplayContext {
 		return new String[] {"create-date"};
 	}
 
+	public String getOrphanPortletsURL(Layout layout) {
+		PortletURL orphanPortletsURL =
+			_liferayPortletResponse.createRenderURL();
+
+		orphanPortletsURL.setParameter("mvcPath", "/orphan_portlets.jsp");
+		orphanPortletsURL.setParameter(
+			"backURL", _themeDisplay.getURLCurrent());
+		orphanPortletsURL.setParameter(
+			"selPlid", String.valueOf(layout.getPlid()));
+
+		return orphanPortletsURL.toString();
+	}
+
 	public String getPagesName() {
 		if (_pagesName != null) {
 			return _pagesName;
@@ -388,6 +464,15 @@ public class LayoutsAdminDisplayContext {
 		}
 
 		return sb.toString();
+	}
+
+	public String getPermissionsURL(Layout layout) throws Exception {
+		return PermissionsURLTag.doTag(
+			StringPool.BLANK, Layout.class.getName(),
+			HtmlUtil.escape(layout.getName(_themeDisplay.getLocale())), null,
+			String.valueOf(layout.getPlid()),
+			LiferayWindowState.POP_UP.toString(), null,
+			_themeDisplay.getRequest());
 	}
 
 	public PortletURL getPortletURL() {
@@ -516,8 +601,18 @@ public class LayoutsAdminDisplayContext {
 			return _millerColumnsEnabled;
 		}
 
-		_millerColumnsEnabled = ParamUtil.getBoolean(
-			_liferayPortletRequest, "millerColumns");
+		_millerColumnsEnabled = false;
+
+		LayoutAdminWebConfiguration layoutAdminWebConfiguration =
+			(LayoutAdminWebConfiguration)_liferayPortletRequest.getAttribute(
+				LayoutAdminWebKeys.LAYOUT_ADMIN_CONFIGURATION);
+
+		if (Objects.equals(
+				layoutAdminWebConfiguration.layoutDisplayStyle(),
+				LayoutAdminDisplayStyleKeys.MILLER)) {
+
+			_millerColumnsEnabled = true;
+		}
 
 		return _millerColumnsEnabled;
 	}
@@ -591,6 +686,19 @@ public class LayoutsAdminDisplayContext {
 		return true;
 	}
 
+	public boolean showAddChildPageAction(Layout layout)
+		throws PortalException {
+
+		return LayoutPermissionUtil.contains(
+			_themeDisplay.getPermissionChecker(), layout,
+			ActionKeys.ADD_LAYOUT);
+	}
+
+	public boolean showConfigureAction(Layout layout) throws PortalException {
+		return LayoutPermissionUtil.contains(
+			_themeDisplay.getPermissionChecker(), layout, ActionKeys.UPDATE);
+	}
+
 	public boolean showCopyApplicationsAction(Layout layout)
 		throws PortalException {
 
@@ -623,6 +731,106 @@ public class LayoutsAdminDisplayContext {
 			_themeDisplay.getPermissionChecker(), layout, ActionKeys.UPDATE);
 	}
 
+	public boolean showDeleteAction(Layout layout) throws PortalException {
+		if (StagingUtil.isIncomplete(layout)) {
+			return false;
+		}
+
+		if (!LayoutPermissionUtil.contains(
+				_themeDisplay.getPermissionChecker(), layout,
+				ActionKeys.DELETE)) {
+
+			return false;
+		}
+
+		Group group = layout.getGroup();
+
+		int layoutsCount = LayoutLocalServiceUtil.getLayoutsCount(
+			group, false, LayoutConstants.DEFAULT_PARENT_LAYOUT_ID);
+
+		if (group.isGuest() && !layout.isPrivateLayout() &&
+			layout.isRootLayout() && (layoutsCount == 1)) {
+
+			return false;
+		}
+
+		return true;
+	}
+
+	public boolean showOrphanPortletsAction(Layout layout) {
+		if (StagingUtil.isIncomplete(layout)) {
+			return false;
+		}
+
+		if (!layout.isSupportsEmbeddedPortlets()) {
+			return false;
+		}
+
+		OrphanPortletsDisplayContext orphanPortletsDisplayContext =
+			new OrphanPortletsDisplayContext(_liferayPortletRequest);
+
+		if (ListUtil.isEmpty(
+				orphanPortletsDisplayContext.getOrphanPortlets(layout))) {
+
+			return false;
+		}
+
+		return true;
+	}
+
+	public boolean showPermissionsAction(Layout layout) throws PortalException {
+		if (StagingUtil.isIncomplete(layout)) {
+			return false;
+		}
+
+		Group selGroup = getSelGroup();
+
+		if (selGroup.isLayoutPrototype()) {
+			return false;
+		}
+
+		return LayoutPermissionUtil.contains(
+			_themeDisplay.getPermissionChecker(), layout,
+			ActionKeys.PERMISSIONS);
+	}
+
+	private JSONObject _getActionURLsJSONObject(Layout layout)
+		throws Exception {
+
+		JSONObject jsonObject = JSONFactoryUtil.createJSONObject();
+
+		if (showAddChildPageAction(layout)) {
+			jsonObject.put(
+				"addURL",
+				getAddLayoutURL(layout.getPlid(), layout.isPrivateLayout()));
+		}
+
+		if (showConfigureAction(layout)) {
+			jsonObject.put("configureURL", getConfigureLayoutURL(layout));
+		}
+
+		if (showCopyApplicationsAction(layout)) {
+			jsonObject.put(
+				"copyApplicationsURL", getCopyApplicationsURL(layout));
+		}
+
+		if (showDeleteAction(layout)) {
+			jsonObject.put("deleteURL", getDeleteLayoutURL(layout));
+		}
+
+		jsonObject.put("editLayoutURL", getEditLayoutURL(layout));
+
+		if (showOrphanPortletsAction(layout)) {
+			jsonObject.put("orphanPortletsURL", getOrphanPortletsURL(layout));
+		}
+
+		if (showPermissionsAction(layout)) {
+			jsonObject.put("permissionsURL", getPermissionsURL(layout));
+		}
+
+		return jsonObject;
+	}
+
 	private JSONObject _getBreadcrumbEntryJSONObject(long plid, String title)
 		throws PortalException {
 
@@ -652,6 +860,8 @@ public class LayoutsAdminDisplayContext {
 		for (Layout layout : layouts) {
 			JSONObject layoutJSONObject = JSONFactoryUtil.createJSONObject();
 
+			layoutJSONObject.put(
+				"actionURLs", _getActionURLsJSONObject(layout));
 			layoutJSONObject.put("active", _isActive(layout.getPlid()));
 
 			int childLayoutsCount = LayoutLocalServiceUtil.getLayoutsCount(
@@ -701,6 +911,10 @@ public class LayoutsAdminDisplayContext {
 		}
 
 		Layout selLayout = getSelLayout();
+
+		if (selLayout == null) {
+			return false;
+		}
 
 		for (Layout layout : selLayout.getAncestors()) {
 			if (plid == layout.getPlid()) {