+++ modules/apps/web-experience/layout/layout-admin-web/src/main/resources/META-INF/resources/js/Layout.es.js
@@ -49,7 +49,10 @@ class Layout extends Component {
 	 * @inheritDoc
 	 */
 	rendered() {
-		this.refs.layout.scrollLeft = this.refs.layout.scrollWidth;
+		requestAnimationFrame(() => {
+			this.refs.layoutColumns.scrollLeft =
+				this.refs.layoutColumns.scrollWidth;
+		});
 	}
 }
 
@@ -81,6 +84,7 @@ Layout.STATE = {
 	layoutColumns: Config.arrayOf(
 		Config.arrayOf(
 			Config.shapeOf({
+				actionURLs: Config.object().required(),
 				active: Config.bool().required(),
 				hasChild: Config.bool().required(),
 				plid: Config.string().required(),